nginx: set fixed server_names_hash_bucket_size
The value determines how many vhosts the nginx process
can have, so it's important it's not too small.

By default, nginx sets this to a value that depends on
the CPU's cache line width. We don't want to be too sure
that all solarnet machines have the same CPU, so we make
the value the same everywhere.

License: MIT
Signed-off-by: Lars Gierth <larsg@systemli.org>