metrics: trigger prometheus upgrades using ref-file
Changes to the Dockerfile often require a rebuild of the image.
We can't delete the image to trigger a rebuild, because we don't
wanna kill off the Prometheus daemon.

We introduce a ref-file, similar to how cjdns upgrades are handled.
The ref-file contains the currently running git ref (commit, branch,
or tag). We build unless the file exists and contains the expected ref.

In order to trigger a rebuild, simply remove the ref-file:

    ssh root@host rm /opt/prometheus.ref

License: MIT
Signed-off-by: Lars Gierth <larsg@systemli.org>