cjdns: fix installation of cjdroute and upstart script
This bug is really silly: the copy step grabs the file from
the *local* disk, not the actual host's disk. That worked
purely out of incident.

It obviously never installed the correct version of cjdns.

We now install the correct cjdroute, as well as the correct
upstart script. We also restart cjdroute in case the upstart
script changed.

License: MIT
Signed-off-by: Lars Gierth <larsg@systemli.org>