gateway: redirect to HTTPS and fix HSTS header
Except for requests for deeplinks like /ipfs, /ipns, and /api.
We don't want to break existing clients that forgot to use HTTPS.
This definitely isn't the last word on redirecting deeplinks.

We also stop sending the HSTS header over plaintext HTTP,
and start sending it on vhosts that aren't gateway.ipfs.io or ipfs.io.

These HTTPS redirects only happen for hosts that we have certs for.